requirements.txt
Flask
waitress
tensorflow
numpy
Pillow